move bls signer to there own place
