+++ ts/mcl.ts
@@ -31,40 +31,6 @@ export interface keyPair {
 
 export type Domain = Uint8Array;
 
-export interface BlsSignerInterface {
-    pubkey: solG2;
-    sign(message: string): Signature;
-}
-
-export class NullBlsSinger implements BlsSignerInterface {
-    get pubkey(): solG2 {
-        throw new Error("NullBlsSinger has no public key");
-    }
-    sign(message: string): Signature {
-        throw new Error("NullBlsSinger dosen't sign");
-    }
-}
-
-export class BlsSigner implements BlsSignerInterface {
-    static new(domain: Domain) {
-        const keyPair = newKeyPair();
-        return new BlsSigner(domain, keyPair.secret);
-    }
-    private _pubkey: mclG2;
-    constructor(public domain: Domain, private secret: SecretKey) {
-        this._pubkey = getPubkey(secret);
-    }
-    get pubkey(): solG2 {
-        return g2ToHex(this._pubkey);
-    }
-
-    public sign(message: string): Signature {
-        const { signature } = sign(message, this.secret, this.domain);
-        const sol = g1ToHex(signature);
-        return { mcl: signature, sol };
-    }
-}
-
 export function aggregate(signatures: Signature[]): Signature {
     const aggregated = aggreagateRaw(signatures.map(s => s.mcl));
     return { mcl: aggregated, sol: g1ToHex(aggregated) };
